fix: aabb merge

+++ src/intersection.rs
@@ -2,6 +2,6 @@ use vector::Vector;
 
 pub struct Intersection {
   pub position: Vector,
-  pub distance: f64,
+  pub distance: f32,
   pub normal: Vector,
 }
